+++ tests/conftest.py
@@ -1,15 +1,20 @@
+from typing import Any, Dict
+
 from pytest import Session
 
 
 def _patch_plotly_show() -> None:
     """Monkey patch ``plotly.io.show`` as to not perform any rendering and,
     instead, simply call ``plotly.io._utils.validate_coerce_fig_to_dict``"""
+    from typing import Union
+
     import plotly.io
+    from plotly.graph_objs import Figure
+    from plotly.io._utils import validate_coerce_fig_to_dict  # noqa
 
-    # noinspection PyProtectedMember
-    from plotly.io._utils import validate_coerce_fig_to_dict
-
-    def wrapped(fig, renderer=None, validate=True, **kwargs):
+    def wrapped(
+        fig: Union[Figure, dict], renderer: str = None, validate: bool = True, **kwargs: Dict[str, Any]
+    ) -> None:
         validate_coerce_fig_to_dict(fig, validate)
 
     plotly.io.show = wrapped